[Separation of Concerns] atom
