+++ src/components/Atom/BaseLinkButton/index.vue
@@ -1,3 +1,5 @@
+<script src="./index.js"></script>
+<style lang="sass" src="./index.sass" scoped></style>
 <template lang="pug">
   a(
     target="_blank"
@@ -8,34 +10,3 @@
     slot
     span(v-text="text")
 </template>
-
-<script>
-export default {
-  name: 'BaseLinkButton',
-  props: {
-    colorType: {
-      type: String,
-      default: 'green',
-    },
-    text: {
-      type: String,
-      require: true,
-      default: 'Text',
-    },
-    url: {
-      type: String,
-      default: '###',
-    },
-  },
-  computed: {
-    className() {
-      const rawName = this.colorType
-      const lowcase = rawName.toLowerCase()
-      const upperCamelCase = lowcase.charAt(0).toUpperCase() + lowcase.slice(1)
-      return `button${upperCamelCase}`
-    },
-  },
-}
-</script>
-
-<style lang="sass" src="./index.sass" scoped />
