+++ src/components/Atom/BaseLinkText/index.vue
@@ -1,3 +1,5 @@
+<script src="./index.js"></script>
+<style lang="sass" src="./index.sass" module></style>
 <template lang="pug">
   a(
     target="_blank"
@@ -8,35 +10,3 @@
     slot
     span(v-text="text")
 </template>
-
-<script>
-export default {
-  name: 'BaseLinkText',
-  props: {
-    colorType: {
-      type: String,
-      default: 'green',
-      validator(value) {
-        return ['green', 'grey', 'white'].includes(value)
-      },
-    },
-    text: {
-      type: String,
-      require: true,
-      default: 'Text',
-    },
-    url: {
-      type: String,
-      default: '###',
-    },
-  },
-  computed: {
-    className() {
-      const lowcase = this.colorType.toLowerCase()
-      return this.$style[lowcase]
-    },
-  },
-}
-</script>
-
-<style lang="sass" src="./index.sass" module />
