+++ src/components/Atom/BaseNuxtLink/index.vue
@@ -1,3 +1,5 @@
+<script src="./index.js"></script>
+<style lang="sass" src="./index.sass" module></style>
 <template lang="pug">
   NuxtLink(
     :to="url"
@@ -7,39 +9,3 @@
     slot
     span(v-text="text")
 </template>
-
-<script>
-export default {
-  name: 'BaseNuxtLink',
-  props: {
-    colorType: {
-      type: String,
-      default: 'green',
-      validator(value) {
-        return ['green', 'grey', 'white'].includes(value)
-      },
-    },
-    exactActiveClass: {
-      type: String,
-      default: '',
-    },
-    text: {
-      type: String,
-      require: true,
-      default: 'Text',
-    },
-    url: {
-      type: String,
-      default: '###',
-    },
-  },
-  computed: {
-    className() {
-      const lowcase = this.colorType.toLowerCase()
-      return this.$style[lowcase]
-    },
-  },
-}
-</script>
-
-<style lang="sass" src="./index.sass" module />
