+++ src/components/Atom/Typography/index.js
@@ -1,0 +1,21 @@
+export default {
+  name: 'Typography',
+  props: {
+    text: {
+      type: String,
+      default: '15:00',
+    },
+    level: {
+      type: Number,
+      default: 2,
+      validator(value) {
+        return value >= 1 && value <= 4
+      },
+    },
+  },
+  computed: {
+    tagName() {
+      return `h${this.level}`
+    },
+  },
+}
