+++ src/components/Atom/Typography/index.vue
@@ -1,3 +1,5 @@
+<script src="./index.js"></script>
+<style lang="sass" src="./index.sass" scoped></style>
 <template lang="pug">
   component(
     :is="tagName"
@@ -5,29 +7,3 @@
     v-text="text"
   )
 </template>
-
-<script>
-export default {
-  name: 'Typography',
-  props: {
-    text: {
-      type: String,
-      default: '15:00',
-    },
-    level: {
-      type: Number,
-      default: 2,
-      validator(value) {
-        return value >= 1 && value <= 4
-      },
-    },
-  },
-  computed: {
-    tagName() {
-      return `h${this.level}`
-    },
-  },
-}
-</script>
-
-<style lang="sass" src="./index.sass" scoped />
