+++ src/components/Atom/VueToNuxtLogo/index.js
@@ -1,0 +1,3 @@
+export default {
+  name: 'VueToNuxtLogo',
+}
