+++ src/components/Atom/VueToNuxtLogo/index.vue
@@ -1,3 +1,5 @@
+<script src="./index.js"></script>
+<style lang="sass" src="./index.sass" module></style>
 <template lang="pug">
   div(:class="$style.vueToNuxtLogo")
     div(:class="$style.two")
@@ -5,11 +7,3 @@
     div(:class="$style.three")
     div(:class="$style.four")
 </template>
-
-<script>
-export default {
-  name: 'VueToNuxtLogo',
-}
-</script>
-
-<style lang="sass" src="./index.sass" module />
